docker/face-api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
numpy
Pillow
opencv-python-headless
ImageHash
